requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
dnspython==2.1.0
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.3
Werkzeug==1.0.1
